style: fix mpp lint issues

+++ sample/sample_uvc_vdec_venc.c
@@ -41,30 +41,28 @@
 
 /* ======================== Config ======================== */
 
-#define SAMPLE_WIDTH        1280
-#define SAMPLE_HEIGHT       720
-#define SAMPLE_FPS          30
-#define SAMPLE_BITRATE      2000000     /* 2 Mbps */
-#define SAMPLE_GOP          30
-#define SAMPLE_SAVE_COUNT   60          /* encoded frames to save */
-#define SAMPLE_WARMUP_COUNT 2           /* warm-up frames to discard (manual mode) */
+#define SAMPLE_WIDTH 1280
+#define SAMPLE_HEIGHT 720
+#define SAMPLE_FPS 30
+#define SAMPLE_BITRATE 2000000 /* 2 Mbps */
+#define SAMPLE_GOP 30
+#define SAMPLE_SAVE_COUNT 60  /* encoded frames to save */
+#define SAMPLE_WARMUP_COUNT 2 /* warm-up frames to discard (manual mode) */
 
-static const char *g_devNode  = "/dev/video13";
-static const char *g_outFile  = "./output.h264";
-static S32         g_bindMode = 0;
+static const char *g_devNode = "/dev/video13";
+static const char *g_outFile = "./output.h264";
+static S32 g_bindMode = 0;
 
 static volatile S32 g_running = 1;
 
-static void sig_handler(int sig)
-{
+static void sig_handler(int sig) {
     (void)sig;
     g_running = 0;
 }
 
 /* ======================== Write H.264 Stream ======================== */
 
-static S32 write_stream(FILE *fp, const StreamBufferInfo *pstStream, U32 u32Idx)
-{
+static S32 write_stream(FILE *fp, const StreamBufferInfo *pstStream, U32 u32Idx) {
     if (!pstStream->pu8Addr || pstStream->u32Size == 0)
         return -1;
 
@@ -74,16 +72,18 @@ static S32 write_stream(FILE *fp, const StreamBufferInfo *pstStream, U32 u32Idx)
         return -1;
     }
 
-    printf("  [SAVE] frame %04u: %u bytes  key=%d  pts=%llu\n",
-           u32Idx, pstStream->u32Size, pstStream->bKeyFrame,
-           (unsigned long long)pstStream->u64PTS);
+    printf(
+        "  [SAVE] frame %04u: %u bytes  key=%d  pts=%" PRIu64 "\n",
+        u32Idx,
+        pstStream->u32Size,
+        pstStream->bKeyFrame,
+        (uint64_t)pstStream->u64PTS);
     return 0;
 }
 
 /* ======================== Manual Pipeline ======================== */
 
-static S32 run_manual(void)
-{
+static S32 run_manual(void) {
     S32 ret;
 
     if (access(g_devNode, F_OK) != 0) {
@@ -124,11 +124,11 @@ static S32 run_manual(void)
     UVC_CHN uvcChn = 0;
     UvcChnAttr uvcChnAttr;
     memset(&uvcChnAttr, 0, sizeof(uvcChnAttr));
-    uvcChnAttr.u32Width     = SAMPLE_WIDTH;
-    uvcChnAttr.u32Height    = SAMPLE_HEIGHT;
+    uvcChnAttr.u32Width = SAMPLE_WIDTH;
+    uvcChnAttr.u32Height = SAMPLE_HEIGHT;
     uvcChnAttr.ePixelFormat = MPP_PIXEL_FORMAT_MJPEG;
-    uvcChnAttr.u32Fps       = SAMPLE_FPS;
-    uvcChnAttr.u32Depth     = 1;
+    uvcChnAttr.u32Fps = SAMPLE_FPS;
+    uvcChnAttr.u32Depth = 1;
 
     ret = UVC_SetChnAttr(uvcDev, uvcChn, &uvcChnAttr);
     assert(ret == 0);
@@ -143,10 +143,10 @@ static S32 run_manual(void)
     S32 vdecChn = 0;
     VdecChnAttr vdecAttr;
     memset(&vdecAttr, 0, sizeof(vdecAttr));
-    vdecAttr.eCodecType         = MPP_STREAM_CODEC_MJPEG;
+    vdecAttr.eCodecType = MPP_STREAM_CODEC_MJPEG;
     vdecAttr.eOutputPixelFormat = MPP_PIXEL_FORMAT_NV12;
-    vdecAttr.u32Width           = SAMPLE_WIDTH;
-    vdecAttr.u32Height          = SAMPLE_HEIGHT;
+    vdecAttr.u32Width = SAMPLE_WIDTH;
+    vdecAttr.u32Height = SAMPLE_HEIGHT;
 
     ret = VDEC_CreateChn(vdecChn, &vdecAttr);
     assert(ret == 0);
@@ -162,14 +162,14 @@ static S32 run_manual(void)
     S32 vencChn = 0;
     VencChnAttr vencAttr;
     memset(&vencAttr, 0, sizeof(vencAttr));
-    vencAttr.eCodecType        = MPP_STREAM_CODEC_H264;
+    vencAttr.eCodecType = MPP_STREAM_CODEC_H264;
     vencAttr.eInputPixelFormat = MPP_PIXEL_FORMAT_NV12;
-    vencAttr.u32Width          = SAMPLE_WIDTH;
-    vencAttr.u32Height         = SAMPLE_HEIGHT;
-    vencAttr.u32Bitrate        = SAMPLE_BITRATE;
-    vencAttr.u32FrameRate      = SAMPLE_FPS;
-    vencAttr.u32Gop            = SAMPLE_GOP;
-    vencAttr.eRcMode           = VENC_RC_MODE_CBR;
+    vencAttr.u32Width = SAMPLE_WIDTH;
+    vencAttr.u32Height = SAMPLE_HEIGHT;
+    vencAttr.u32Bitrate = SAMPLE_BITRATE;
+    vencAttr.u32FrameRate = SAMPLE_FPS;
+    vencAttr.u32Gop = SAMPLE_GOP;
+    vencAttr.eRcMode = VENC_RC_MODE_CBR;
 
     ret = VENC_CreateChn(vencChn, &vencAttr);
     assert(ret == 0);
@@ -193,21 +193,19 @@ static S32 run_manual(void)
 
     /* --- Main loop: UVC → VDEC → VENC → file --- */
     U32 u32Saved = 0;
-    printf("  [INFO] Pipeline running (manual), saving %u frames to %s\n",
-           SAMPLE_SAVE_COUNT, g_outFile);
+    printf("  [INFO] Pipeline running (manual), saving %u frames to %s\n", SAMPLE_SAVE_COUNT, g_outFile);
 
     while (g_running && u32Saved < SAMPLE_SAVE_COUNT) {
-
         /* 1. Get MJPEG frame from UVC */
         memset(&uvcFrame, 0, sizeof(uvcFrame));
         ret = UVC_GetFrame(uvcDev, uvcChn, &uvcFrame, 3000);
         if (ret != 0) {
-            if (!g_running) break;
+            if (!g_running)
+                break;
             continue;
         }
 
-        if (uvcFrame.stVFrame.ulPlaneVirAddr[0] == 0 ||
-            uvcFrame.stVFrame.u32PlaneSizeValid[0] == 0) {
+        if (uvcFrame.stVFrame.ulPlaneVirAddr[0] == 0 || uvcFrame.stVFrame.u32PlaneSizeValid[0] == 0) {
             UVC_ReleaseFrame(uvcDev, uvcChn, &uvcFrame);
             continue;
         }
@@ -215,12 +213,12 @@ static S32 run_manual(void)
         /* 2. Send MJPEG to VDEC */
         StreamBufferInfo stream;
         memset(&stream, 0, sizeof(stream));
-        stream.pu8Addr      = (const U8 *)uvcFrame.stVFrame.ulPlaneVirAddr[0];
-        stream.u32Size      = uvcFrame.stVFrame.u32PlaneSizeValid[0];
-        stream.eCodecType   = MPP_STREAM_CODEC_MJPEG;
-        stream.bKeyFrame    = MPP_TRUE;
+        stream.pu8Addr = (const U8 *)uvcFrame.stVFrame.ulPlaneVirAddr[0];
+        stream.u32Size = uvcFrame.stVFrame.u32PlaneSizeValid[0];
+        stream.eCodecType = MPP_STREAM_CODEC_MJPEG;
+        stream.bKeyFrame = MPP_TRUE;
         stream.bEndOfStream = MPP_FALSE;
-        stream.u64PTS       = uvcFrame.stVFrame.u64PTS;
+        stream.u64PTS = uvcFrame.stVFrame.u64PTS;
 
         ret = VDEC_SendStream(vdecChn, &stream, 0);
         printf("  [DBG ] VDEC_SendStream ret=%d size=%u\n", ret, stream.u32Size);
@@ -233,15 +231,16 @@ static S32 run_manual(void)
         VideoFrameInfo decFrame;
         memset(&decFrame, 0, sizeof(decFrame));
         ret = VDEC_GetFrame(vdecChn, &decFrame, 1000);
-        printf("  [DBG ] VDEC_GetFrame ret=%d bufferId=%lu w=%u h=%u planes=%u"
-               " virAddr[0]=%lu sizeValid[0]=%u\n",
-               ret,
-               (unsigned long)decFrame.ulBufferId,
-               decFrame.stVdecFrameInfo.stCommFrameInfo.u32Width,
-               decFrame.stVdecFrameInfo.stCommFrameInfo.u32Height,
-               decFrame.stVFrame.u32PlaneNum,
-               (unsigned long)decFrame.stVFrame.ulPlaneVirAddr[0],
-               decFrame.stVFrame.u32PlaneSizeValid[0]);
+        printf(
+            "  [DBG ] VDEC_GetFrame ret=%d bufferId=%lu w=%u h=%u planes=%u"
+            " virAddr[0]=%lu sizeValid[0]=%u\n",
+            ret,
+            (uintptr_t)decFrame.ulBufferId,
+            decFrame.stVdecFrameInfo.stCommFrameInfo.u32Width,
+            decFrame.stVdecFrameInfo.stCommFrameInfo.u32Height,
+            decFrame.stVFrame.u32PlaneNum,
+            (uintptr_t)decFrame.stVFrame.ulPlaneVirAddr[0],
+            decFrame.stVFrame.u32PlaneSizeValid[0]);
         if (ret == ERR_VDEC_NO_FRAME || ret == ERR_VDEC_TIMEOUT)
             continue;
         if (ret == ERR_VDEC_EOS)
@@ -253,7 +252,7 @@ static S32 run_manual(void)
 
         /* 4. Send NV12 frame to VENC */
         ret = VENC_SendFrame(vencChn, &decFrame, 0);
-        printf("  [DBG ] VENC_SendFrame ret=%d bufferId=%lu\n", ret, (unsigned long)decFrame.ulBufferId);
+        printf("  [DBG ] VENC_SendFrame ret=%d bufferId=%lu\n", ret, (uintptr_t)decFrame.ulBufferId);
 
         /* Release decoded frame after VENC has consumed it */
         VDEC_ReleaseFrame(vdecChn, decFrame.ulBufferId);
@@ -311,8 +310,7 @@ manual_cleanup_uvc_dev:
 
 /* ======================== Bind Pipeline ======================== */
 
-static S32 run_bind(void)
-{
+static S32 run_bind(void) {
     S32 ret;
 
     if (access(g_devNode, F_OK) != 0) {
@@ -353,11 +351,11 @@ static S32 run_bind(void)
     UVC_CHN uvcChn = 0;
     UvcChnAttr uvcChnAttr;
     memset(&uvcChnAttr, 0, sizeof(uvcChnAttr));
-    uvcChnAttr.u32Width     = SAMPLE_WIDTH;
-    uvcChnAttr.u32Height    = SAMPLE_HEIGHT;
+    uvcChnAttr.u32Width = SAMPLE_WIDTH;
+    uvcChnAttr.u32Height = SAMPLE_HEIGHT;
     uvcChnAttr.ePixelFormat = MPP_PIXEL_FORMAT_MJPEG;
-    uvcChnAttr.u32Fps       = SAMPLE_FPS;
-    uvcChnAttr.u32Depth     = 0;
+    uvcChnAttr.u32Fps = SAMPLE_FPS;
+    uvcChnAttr.u32Depth = 0;
 
     ret = UVC_SetChnAttr(uvcDev, uvcChn, &uvcChnAttr);
     assert(ret == 0);
@@ -372,10 +370,10 @@ static S32 run_bind(void)
     S32 vdecChn = 0;
     VdecChnAttr vdecAttr;
     memset(&vdecAttr, 0, sizeof(vdecAttr));
-    vdecAttr.eCodecType         = MPP_STREAM_CODEC_MJPEG;
+    vdecAttr.eCodecType = MPP_STREAM_CODEC_MJPEG;
     vdecAttr.eOutputPixelFormat = MPP_PIXEL_FORMAT_NV12;
-    vdecAttr.u32Width           = SAMPLE_WIDTH;
-    vdecAttr.u32Height          = SAMPLE_HEIGHT;
+    vdecAttr.u32Width = SAMPLE_WIDTH;
+    vdecAttr.u32Height = SAMPLE_HEIGHT;
 
     ret = VDEC_CreateChn(vdecChn, &vdecAttr);
     assert(ret == 0);
@@ -391,14 +389,14 @@ static S32 run_bind(void)
     S32 vencChn = 0;
     VencChnAttr vencAttr;
     memset(&vencAttr, 0, sizeof(vencAttr));
-    vencAttr.eCodecType        = MPP_STREAM_CODEC_H264;
+    vencAttr.eCodecType = MPP_STREAM_CODEC_H264;
     vencAttr.eInputPixelFormat = MPP_PIXEL_FORMAT_NV12;
-    vencAttr.u32Width          = SAMPLE_WIDTH;
-    vencAttr.u32Height         = SAMPLE_HEIGHT;
-    vencAttr.u32Bitrate        = SAMPLE_BITRATE;
-    vencAttr.u32FrameRate      = SAMPLE_FPS;
-    vencAttr.u32Gop            = SAMPLE_GOP;
-    vencAttr.eRcMode           = VENC_RC_MODE_CBR;
+    vencAttr.u32Width = SAMPLE_WIDTH;
+    vencAttr.u32Height = SAMPLE_HEIGHT;
+    vencAttr.u32Bitrate = SAMPLE_BITRATE;
+    vencAttr.u32FrameRate = SAMPLE_FPS;
+    vencAttr.u32Gop = SAMPLE_GOP;
+    vencAttr.eRcMode = VENC_RC_MODE_CBR;
 
     ret = VENC_CreateChn(vencChn, &vencAttr);
     assert(ret == 0);
@@ -411,18 +409,17 @@ static S32 run_bind(void)
     }
 
     /* --- Bind: UVC → VDEC → VENC --- */
-    MppNode stUvcSrc  = { .eModId = MPP_ID_UVC,  .s32DevId = uvcDev, .s32ChnId = uvcChn };
-    MppNode stVdecSink = { .eModId = MPP_ID_VDEC, .s32DevId = 0,      .s32ChnId = vdecChn };
-    MppNode stVdecSrc  = { .eModId = MPP_ID_VDEC, .s32DevId = 0,      .s32ChnId = vdecChn };
-    MppNode stVencSink = { .eModId = MPP_ID_VENC, .s32DevId = 0,      .s32ChnId = vencChn };
+    MppNode stUvcSrc = {.eModId = MPP_ID_UVC, .s32DevId = uvcDev, .s32ChnId = uvcChn};
+    MppNode stVdecSink = {.eModId = MPP_ID_VDEC, .s32DevId = 0, .s32ChnId = vdecChn};
+    MppNode stVdecSrc = {.eModId = MPP_ID_VDEC, .s32DevId = 0, .s32ChnId = vdecChn};
+    MppNode stVencSink = {.eModId = MPP_ID_VENC, .s32DevId = 0, .s32ChnId = vencChn};
 
     ret = SYS_Bind(&stUvcSrc, &stVdecSink);
     if (ret != 0) {
         printf("  [ERROR] SYS_Bind UVC→VDEC failed (ret=%d)\n", ret);
         goto bind_cleanup_venc;
     }
-    printf("  [INFO] SYS_Bind: UVC(dev=%d,chn=%d) → VDEC(chn=%d) OK\n",
-           uvcDev, uvcChn, vdecChn);
+    printf("  [INFO] SYS_Bind: UVC(dev=%d,chn=%d) → VDEC(chn=%d) OK\n", uvcDev, uvcChn, vdecChn);
 
     ret = SYS_Bind(&stVdecSrc, &stVencSink);
     if (ret != 0) {
@@ -430,13 +427,11 @@ static S32 run_bind(void)
         SYS_UnBind(&stUvcSrc, &stVdecSink);
         goto bind_cleanup_venc;
     }
-    printf("  [INFO] SYS_Bind: VDEC(chn=%d) → VENC(chn=%d) OK\n",
-           vdecChn, vencChn);
+    printf("  [INFO] SYS_Bind: VDEC(chn=%d) → VENC(chn=%d) OK\n", vdecChn, vencChn);
 
     /* --- Main loop: read encoded H.264 from VENC --- */
     U32 u32Saved = 0;
-    printf("  [INFO] Pipeline running (bind), saving %u frames to %s\n",
-           SAMPLE_SAVE_COUNT, g_outFile);
+    printf("  [INFO] Pipeline running (bind), saving %u frames to %s\n", SAMPLE_SAVE_COUNT, g_outFile);
 
     while (g_running && u32Saved < SAMPLE_SAVE_COUNT) {
         StreamBufferInfo encStream;
@@ -494,29 +489,32 @@ bind_cleanup_uvc_dev:
 
 /* ======================== Main ======================== */
 
-int main(int argc, char *argv[])
-{
+int main(int argc, char *argv[]) {
     S32 argIdx = 1;
     while (argIdx < argc && argv[argIdx][0] == '-') {
         if (strcmp(argv[argIdx], "--bind") == 0) {
             g_bindMode = 1;
-        } else if (strcmp(argv[argIdx], "--help") == 0 ||
-                   strcmp(argv[argIdx], "-h") == 0) {
-            printf("Usage: %s [OPTIONS] [devNode] [outFile]\n\n"
-                   "  UVC capture -> VDEC decode -> VENC encode -> save H.264.\n\n"
-                   "Options:\n"
-                   "  --bind    Use SYS_Bind mode (UVC->VDEC->VENC automatic delivery).\n"
-                   "            Default is manual mode.\n"
-                   "  -h,--help Show this help message.\n\n"
-                   "Positional:\n"
-                   "  devNode   UVC device node  (default: %s)\n"
-                   "  outFile   Output H.264 file (default: %s)\n\n"
-                   "Examples:\n"
-                   "  %s\n"
-                   "  %s --bind\n"
-                   "  %s --bind /dev/video0 ./out.h264\n",
-                   argv[0], g_devNode, g_outFile,
-                   argv[0], argv[0], argv[0]);
+        } else if (strcmp(argv[argIdx], "--help") == 0 || strcmp(argv[argIdx], "-h") == 0) {
+            printf(
+                "Usage: %s [OPTIONS] [devNode] [outFile]\n\n"
+                "  UVC capture -> VDEC decode -> VENC encode -> save H.264.\n\n"
+                "Options:\n"
+                "  --bind    Use SYS_Bind mode (UVC->VDEC->VENC automatic delivery).\n"
+                "            Default is manual mode.\n"
+                "  -h,--help Show this help message.\n\n"
+                "Positional:\n"
+                "  devNode   UVC device node  (default: %s)\n"
+                "  outFile   Output H.264 file (default: %s)\n\n"
+                "Examples:\n"
+                "  %s\n"
+                "  %s --bind\n"
+                "  %s --bind /dev/video0 ./out.h264\n",
+                argv[0],
+                g_devNode,
+                g_outFile,
+                argv[0],
+                argv[0],
+                argv[0]);
             return 0;
         }
         argIdx++;
@@ -526,11 +524,10 @@ int main(int argc, char *argv[])
     if (argIdx < argc)
         g_outFile = argv[argIdx++];
 
-    signal(SIGINT,  sig_handler);
+    signal(SIGINT, sig_handler);
     signal(SIGTERM, sig_handler);
 
-    printf("=== Sample: UVC → VDEC → VENC (%s) → H.264 ===\n",
-           g_bindMode ? "Bind" : "Manual");
+    printf("=== Sample: UVC → VDEC → VENC (%s) → H.264 ===\n", g_bindMode ? "Bind" : "Manual");
     printf("  Device : %s\n", g_devNode);
     printf("  Output : %s\n", g_outFile);
     printf("  Size   : %ux%u @ %u fps\n", SAMPLE_WIDTH, SAMPLE_HEIGHT, SAMPLE_FPS);
